+++ changelog.md
@@ -2,9 +2,15 @@
 
 ## Added
 * Creative branch of bees - generate music, books, and spooks!
+* Terraforming branch of bees - change biomes!
+* Biome mutation condition, check the wiki for more info on how to use it
+* A lot of datagen stuff internally
 
 ## Changed
 * A LOT of bees' default genes - generally to balance out some power spikes on the curve to a perfect bee and to encourage breeding into some of the 'specialty' branches to get better genes
 
 ## Fixed
 * Territory gene wasn't correctly serializing
+* Non-overworld dimension sky check working incorrectly
+* Surreal and Enigmatic bees had no flower type
+* Rare issue where several things related to the discovery/research tracker would crash or break on removal/addition of a species